backend/requirements.txt
django-filter==22.1
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-phonenumber-field==5.0.0
phonenumbers==8.12.9
django-rest-passwordreset==1.3.0